API - move `set-admin` command in users blueprint
