+++ docker/set-admin.sh
@@ -4,4 +4,4 @@ cd /usr/src/app
 
 source .env.docker
 
-flask set-admin $1
+flask users set-admin $1
