+++ fittrackee/__main__.py
@@ -3,14 +3,11 @@
 import os
 from typing import Dict, Optional
 
-import click
 import gunicorn.app.base
 from flask import Flask
 from flask_migrate import upgrade
 
 from fittrackee import create_app
-from fittrackee.users.exceptions import UserNotFoundException
-from fittrackee.users.utils import set_admin_rights
 
 HOST = os.getenv('HOST', '0.0.0.0')
 PORT = os.getenv('PORT', '5000')
@@ -43,17 +40,6 @@ class StandaloneApplication(gunicorn.app.base.BaseApplication):
 def upgrade_db() -> None:
     with app.app_context():
         upgrade(directory=BASEDIR + '/migrations')
-
-
-@app.cli.command('set-admin')
-@click.argument('username')
-def set_admin(username: str) -> None:
-    """Set admin rights for given user"""
-    try:
-        set_admin_rights(username)
-        print(f"User '{username}' updated.")
-    except UserNotFoundException:
-        print(f"User '{username}' not found.")
 
 
 def main() -> None:
